pyproject.toml
# Copyright (c) Meta Platforms, Inc. and affiliates.
# All rights reserved.
#
# This source code is licensed under the BSD-style license found in the
# LICENSE file in the root directory of this source tree.

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-safescroll-env"
version = "0.1.0"
description = "SafeScroll — Content Moderation & Child Safety Environment for OpenEnv"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]
baseline = [
    "openai>=1.0.0",
]

[project.scripts]
server = "safescroll_env.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["safescroll_env", "safescroll_env.server"]
package-dir = { "safescroll_env" = ".", "safescroll_env.server" = "server" }
